client: move discovery client into client pkg
There exists a client in the etcd/discovery pkg that wraps a
etcd/client.Client and adds logic specifically for interacting with an
etcd discovery service. Move it into client package because it is a
general client logic.

It does minimal change when move client.
