+++ etcdmain/etcd.go
@@ -27,6 +27,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/coreos/etcd/client"
 	"github.com/coreos/etcd/discovery"
 	"github.com/coreos/etcd/etcdserver"
 	"github.com/coreos/etcd/etcdserver/etcdhttp"
@@ -97,7 +98,7 @@ func Main() {
 	shouldProxy := cfg.isProxy() || which == dirProxy
 	if !shouldProxy {
 		stopped, err = startEtcd(cfg)
-		if err == discovery.ErrFullCluster && cfg.shouldFallbackToProxy() {
+		if err == client.ErrFullCluster && cfg.shouldFallbackToProxy() {
 			plog.Noticef("discovery cluster full, falling back to %s", fallbackFlagProxy)
 			shouldProxy = true
 		}
@@ -107,13 +108,13 @@ func Main() {
 	}
 	if err != nil {
 		switch err {
-		case discovery.ErrDuplicateID:
+		case client.ErrDuplicateID:
 			plog.Errorf("member %q has previously registered with discovery service token (%s).", cfg.name, cfg.durl)
 			plog.Errorf("But etcd could not find valid cluster configuration in the given data dir (%s).", cfg.dir)
 			plog.Infof("Please check the given data dir path if the previous bootstrap succeeded")
 			plog.Infof("or use a new discovery token if the previous bootstrap failed.")
 			os.Exit(1)
-		case discovery.ErrDuplicateName:
+		case client.ErrDuplicateName:
 			plog.Errorf("member with duplicated name has registered with discovery service token(%s).", cfg.durl)
 			plog.Errorf("please check (cURL) the discovery token for more information.")
 			plog.Errorf("please do not reuse the discovery token and generate a new one to bootstrap the cluster.")
